Created conftest.py for testing fixtures
